Fix @appetrosyan remarks
