+++ client/tests/integration/events/data.rs
@@ -122,11 +122,9 @@ fn transaction_execution_should_produce_events(executable: Executable) -> Result
     // assertion
     for i in 0..4_usize {
         let domain_id = DomainId::test(&i.to_string());
-        let expected_event = DomainEvent::OtherDomainChange(OtherDomainChangeEvent::new(
-            domain_id,
-            DataStatus::Created,
-        ))
-        .into();
+        let expected_event =
+            DomainEvent::StatusUpdated(DomainStatusUpdated::new(domain_id, DataStatus::Created))
+                .into();
         let event: DataEvent = event_receiver.recv()??.try_into()?;
         assert_eq!(event, expected_event);
     }
