+++ data_model/src/events/data/mod.rs
@@ -23,9 +23,9 @@ pub mod prelude {
     pub use super::RoleEvent;
     pub use super::{
         events::{
-            AccountEvent, AssetDefinitionEvent, AssetEvent, AssetUpdated, DomainEvent,
-            Event as DataEvent, MetadataUpdated, OtherAccountChangeEvent, OtherDomainChangeEvent,
-            PeerEvent, Status as DataStatus, Updated, WorldEvent,
+            AccountEvent, AccountStatusUpdated, AssetDefinitionEvent, AssetEvent, AssetUpdated,
+            DomainEvent, DomainStatusUpdated, Event as DataEvent, MetadataUpdated, PeerEvent,
+            Status as DataStatus, Updated, WorldEvent,
         },
         filters::{EventFilter as DataEventFilter, FilterOpt::*, *},
     };
